feat: implement macaroon middleware
